improve(auth-context): rename file and update import statements
