+++ client/src/hooks/useNavigateUser.ts
@@ -1,7 +1,7 @@
 import { useEffect, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 
-import { AuthContext } from "@/contexts/AuthContext";
+import { AuthContext } from "@/contexts/auth-context";
 
 const useNavigateUser = () => {
   const navigate = useNavigate();
